--- src/hadeswap/constants.ts
@@ -1,1 +1,0 @@
-export const SOL_WITHDRAW_ORDERS_LIMIT__PER_TXN = 4;
