--- src/hadeswap/hooks/usePoolChange/index.ts
@@ -1,1 +1,0 @@
-export * from './usePoolChange';
